The scripts were fine, must be executed
Aside from an instance where I commented out a part of the code, and
15-17 where I just added print to the last lines, all of the scripts are
fine - it's just that for some reason, ipython has exported them in such
a way that if you copy and paste them, it doesn't get the indentation
right. However, if you execute the scripts directly (like I did with
pypy) they are all fully operational.